change the name of onthispage to tablecontents
